#8 [FIX] hooks 폴더 분리
